Avancements

+++ src/Model/graph/Vertex.java
@@ -14,8 +14,8 @@ public class Vertex {
     private Edge leftEdge = null;
     private Edge rightEdge = null;
     boolean isHover;
-    private int x;
-    private int y;
+    private double x;
+    private double y;
     public Vertex()
     {
         isHover=false;
@@ -97,19 +97,19 @@ public class Vertex {
         return left && right  && up && center;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
